requirements.txt
albumentations==1.2.1
einops==0.4.1
MedPy==0.4.0
numpy==1.21.5
opencv_python_headless==4.5.4.60
Pillow==9.2.0
PyWavelets==1.3.0
scikit_image==0.19.3
scikit_learn==1.1.2
scipy==1.7.3
SimpleITK==2.2.0
skimage==0.0
timm==0.6.7
torch==1.8.0+cu111
torchio==0.18.84
torchvision==0.9.0+cu111
tqdm==4.64.0
visdom==0.1.8.9
